display tower selection in gui
