+++ gui/src/lib/components/sidebar/sidebar.svelte
@@ -1,90 +1,19 @@
-<script context="module" lang="ts">
-    interface Tower {
-        name: string
-    }
-</script>
-
 <script lang="ts">
-    import { startRound } from '$lib/game/command_handlers/start-round'
+    import { onDestroy } from 'svelte'
 
-    import { useTowerDrag } from './use-tower-drag'
+    import { setTowerSelectionContext } from '../tower-selection-context'
+    import Shop from './shop/shop.svelte'
+    import TowerDetails from './tower-details/tower-details.svelte'
 
-    const towers: Tower[] = [
-        {
-            name: '1'
-        },
-        {
-            name: '2'
-        },
-        {
-            name: '3'
-        },
-        {
-            name: '4'
-        }
-    ]
+    const { selection, deleteContext } = setTowerSelectionContext()
 
-    const { isDragging, handleDragStart, handleDrag, handleDragEnd } =
-        useTowerDrag()
+    onDestroy(() => {
+        deleteContext()
+    })
 </script>
 
-<div class="container">
-    <div class="grid">
-        {#each towers as tower}
-            <button
-                class="cell"
-                disabled={$isDragging}
-                draggable="true"
-                on:dragstart={handleDragStart}
-                on:drag={handleDrag}
-                on:dragend={handleDragEnd}
-            >
-                {tower.name}
-            </button>
-        {/each}
-    </div>
-
-    <div class="actions">
-        <button on:click={() => startRound()}>Start Round</button>
-    </div>
-</div>
-
-<style lang="scss">
-    .container {
-        height: 100%;
-
-        background: black;
-    }
-
-    .grid {
-        display: grid;
-        grid-template-columns: 50% 50%;
-    }
-
-    .cell {
-        padding: 1rem;
-
-        text-align: center;
-    }
-
-    // Borders
-    .grid {
-        border: 1px solid white;
-        border-right: 0;
-        border-top: 0;
-
-        .cell {
-            border: 1px solid white;
-            border-left: 0;
-            border-bottom: 0;
-        }
-    }
-
-    .actions {
-        padding-top: 1rem;
-        width: 100%;
-
-        display: flex;
-        justify-content: center;
-    }
-</style>
+{#if $selection}
+    <TowerDetails selection={$selection} />
+{:else}
+    <Shop />
+{/if}
